+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- **core** - Fix type inference in concat (`LdhopQueryBuilder`).
+
 ## [1.0.0-alpha.7] - 2026-01-10
 
 ### Changed